chore: Re-organize the project
