+++ examples/ExampleBase/app/build.gradle
@@ -1,33 +1,31 @@
 plugins {
-    id 'com.android.application'
+    alias(libs.plugins.android.application)
 }
 
 android {
-    namespace 'truman.android.example.examplebase'
+    namespace 'truman.android.example.example_base'
     compileSdk 35
 
     defaultConfig {
-        applicationId "truman.android.example.examplebase"
+        applicationId "truman.android.example.example_base"
         minSdk 29
-        targetSdk 34
+        targetSdk 35
         versionCode 5
         versionName "1.2.0"
     }
 
+    buildFeatures {
+        viewBinding true
+    }
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
-
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_17
         targetCompatibility JavaVersion.VERSION_17
-    }
-
-    buildFeatures {
-        viewBinding true
     }
 
     flavorDimensions "composition"
@@ -53,7 +51,8 @@ android {
 
 dependencies {
 
-    implementation 'androidx.appcompat:appcompat:1.7.0'
-    implementation 'com.google.android.material:material:1.12.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.2.1'
+    implementation libs.appcompat
+    implementation libs.material
+    implementation libs.activity
+    implementation libs.constraintlayout
 }
