+++ examples/ExampleBase/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">ExampleBase</string>
+    <string name="app_name">Example Base</string>
 </resources>
