+++ examples/ExampleBase/build.gradle
@@ -1,4 +1,4 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-id 'com.android.application' version '8.9.1' apply false
+alias(libs.plugins.android.application) apply false
 }
