+++ examples/ExampleBase/gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,4 @@
-#Mon Feb 05 14:35:15 KST 2024
+#Tue Apr 08 23:01:22 KST 2025
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 distributionUrl=https\://services.gradle.org/distributions/gradle-8.11.1-bin.zip
